Integrate with existing terraform
