+++ src/ops/cli/config.py
@@ -113,15 +113,14 @@ class JinjaConfigGenerator(object):
         return yaml.safe_load(rendered)
 
 class ClusterConfigGenerator(object):
-    def __init__(self, console_args, cluster_config_path, command, template):
+    def __init__(self, console_args, cluster_config_path, template):
         self.template = template
         self.cluster_config_path = cluster_config_path
         self.console_args = console_args
-        self.command = command
 
     def get(self):
         if os.path.isdir(self.cluster_config_path):
-            return {"cluster": None, "inventory": None, "runner_version": "v2"}
+            return {"cluster": None, "inventory": None}
 
         data_loader = DataLoader()
         # data_loader.set_vault_password('627VR8*;YU99B')
