--- tests/e2e/fixture/ee/data/env=dev/cluster=cluster1/conf.yaml
@@ -1,21 +1,0 @@
-infrastructure:
-  ee-k8s-infra:
-    commit: master
-
-account:
-  cloud_provider:
-    aws:
-      profile: test_profile
-
-env:
-  name: dev
-
-region:
-  location: us-east-1
-  name: va6
-
-project:
-  prefix: ee
-
-cluster:
-  name: cluster1
